+++ kafka-reactive-system/src/test/scala/org/patricknoir/kafka/reactive/client/integration/BaseISpec.scala
@@ -77,27 +77,3 @@ object BaseISpec {
                      """.stripMargin
 }
 
-class SimpleAtLeastOnceISpec extends BaseISpec {
-
-  def is = s2"""
-
-    simple test with at least once semantic   $runTest
-
-  """
-
-  def runTest = {
-    startKafka()
-    startAtLeastOnceServer(ServiceCatalog.echo)
-
-    implicit val timeout = Timeout(10 seconds)
-    val client = new KafkaReactiveClient(KafkaRClientSettings.default)
-
-    val fResponse = client.request[String, String]("kafka:echoInbound/echo", "patrick")
-
-    val result: String = Await.result(fResponse, Duration.Inf)
-    stopKafka()
-
-    result must be_==("patrick")
-  }
-
-}
